@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
     <ApplicationIcon>ColoryrWork.ico</ApplicationIcon>
     <MinimumPlatformVersion>7.0</MinimumPlatformVersion>
@@ -18,10 +18,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AvalonEdit" Version="6.1.3.50" />
+    <PackageReference Include="AvalonEdit" Version="6.3.0.90" />
     <PackageReference Include="DiffPlex" Version="1.7.1" />
-    <PackageReference Include="DiffPlex.Wpf" Version="1.3.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
+    <PackageReference Include="DiffPlex.Wpf" Version="1.4.0" />
+    <!--<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />-->
     <PackageReference Include="NotifyIconWpf" Version="1.0.0" />
     <PackageReference Include="SharpZipLib" Version="1.4.2" />
   </ItemGroup>
