@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <BaseOutputPath>..\..\build_out\NetCoreServer</BaseOutputPath>
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="NetCoreServer" Version="6.6.0" />
+    <PackageReference Include="NetCoreServer" Version="8.0.3" />
   </ItemGroup>
 
   <ItemGroup>
