@@ -4,48 +4,50 @@ using ColoryrServer.Core.FileSystem;
 using ColoryrServer.Core.Managers;
 using ColoryrServer.Core.Utils;
 using ColoryrWork.Lib.Build.Object;
-using Newtonsoft.Json;
+using ColoryrWork.Lib.Build;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Text;
 
 namespace ColoryrServer.Core.BuilderPost;
 
-internal static class PostBuildWebSocket
+internal static class BuildSocket
 {
     public static ReMessage Add(BuildOBJ json)
     {
-        if (CodeManager.GetWebSocket(json.UUID) != null)
+        if (CodeManager.GetSocket(json.UUID) != null)
             return new ReMessage
             {
                 Build = false,
-                Message = $"WebSocket[{json.UUID}]已存在"
+                Message = $"Socket[{json.UUID}]已存在"
             };
-        ServerMain.LogOut($"[{json.User}]正在创建WebSocket[{json.UUID}]");
+        ServerMain.LogOut($"[{json.User}]正在创建Socket[{json.UUID}]");
         var time = string.Format("{0:s}", DateTime.Now);
+        string code = Encoding.UTF8.GetString(json.Code.ToLower() == "true" ? DemoResource.Netty : DemoResource.Socket);
         CSFileCode obj = new()
         {
             UUID = json.UUID,
-            Type = CodeType.WebSocket,
+            Type = CodeType.Socket,
             CreateTime = time,
-            Code = DemoResource.WebSocket
+            Code = code
             .Replace(CodeDemo.Name, json.UUID)
         };
-        CodeManager.StorageWebSocket(obj, json.User);
-        GenWebSocket.StartGen(obj, json.User);
-        ServerMain.LogOut($"[{json.User}]创建WebSocket[{json.UUID}]完成");
+        CodeManager.StorageSocket(obj, json.User);
+        GenSocket.StartGen(obj, json.User);
+        ServerMain.LogOut($"[{json.User}]创建Socket[{json.UUID}]完成");
 
         return new ReMessage
         {
             Build = true,
-            Message = $"WebSocket[{json.UUID}]已创建"
+            Message = $"Socket[{json.UUID}]已创建"
         };
     }
 
     public static CSFileList GetList()
     {
         var list = new CSFileList();
-        foreach (var item in CodeManager.WebSocketFileList)
+        foreach (var item in CodeManager.SocketFileList)
         {
             list.List.Add(item.Key, item.Value);
         }
@@ -55,25 +57,25 @@ internal static class PostBuildWebSocket
 
     public static ReMessage Remove(BuildOBJ json)
     {
-        CodeManager.RemoveFile(CodeType.WebSocket, json.UUID, json.User);
-        AssemblyList.RemoveWebSocket(json.UUID);
+        CodeManager.RemoveFile(CodeType.Socket, json.UUID, json.User);
+        AssemblyList.RemoveSocket(json.UUID);
 
         return new ReMessage
         {
             Build = true,
-            Message = $"WebSocket[{json.UUID}]已删除"
+            Message = $"Socket[{json.UUID}]已删除"
         };
     }
 
     public static ReMessage Updata(BuildOBJ json)
     {
-        var obj = CodeManager.GetWebSocket(json.UUID);
+        var obj = CodeManager.GetSocket(json.UUID);
         if (obj == null)
         {
             return new ReMessage
             {
                 Build = false,
-                Message = $"没有这个WebSocket[{json.UUID}]"
+                Message = $"没有这个Socket[{json.UUID}]"
             };
         }
         if (obj.Version != json.Version)
@@ -81,17 +83,17 @@ internal static class PostBuildWebSocket
             return new ReMessage
             {
                 Build = false,
-                Message = $"WebSocket[{json.UUID}]版本号错误"
+                Message = $"Socket[{json.UUID}]版本号错误"
             };
         }
 
-        var list = JsonConvert.DeserializeObject<List<CodeEditObj>>(json.Code);
+        var list = JsonUtils.ToObj<List<CodeEditObj>>(json.Code);
         obj.Code = FileUtils.StartEdit(obj.Code, list);
         obj.Text = json.Text;
 
         var sw = new Stopwatch();
         sw.Start();
-        var build = GenWebSocket.StartGen(obj, json.User);
+        var build = GenSocket.StartGen(obj, json.User);
         sw.Stop();
         obj.Up();
         return new ReMessage
