@@ -61,56 +61,22 @@ namespace ColoryrServer.Core {
         }
         
         /// <summary>
-        ///   查找类似 &lt;!DOCTYPE html&gt;
-        ///&lt;html&gt;
-        ///
-        ///&lt;head&gt;
-        ///    &lt;meta charset=&quot;utf-8&quot;&gt;
-        ///    &lt;title&gt;测试&lt;/title&gt;
-        ///&lt;/head&gt;
-        ///&lt;script src=&quot;https://cdn.jsdelivr.net/npm/vue/dist/vue.js&quot;&gt;&lt;/script&gt;
-        ///&lt;script src=&quot;https://unpkg.com/axios/dist/axios.min.js&quot;&gt;&lt;/script&gt;
-        ///
-        ///&lt;body&gt;
-        ///    &lt;div id=&quot;app&quot;&gt;
-        ///        &lt;h1&gt;Bitcoin Price Index&lt;/h1&gt;
-        ///        &lt;section v-if=&quot;errored&quot;&gt;
-        ///            &lt;p&gt;We&apos;re sorry, we&apos;re not able to retrieve this information at the moment, please try back later&lt;/p&gt;
-        ///        &lt;/section&gt;
-        ///        &lt;section v-else&gt;
-        ///            &lt;div v [字符串的其余部分被截断]&quot;; 的本地化字符串。
+        ///   查找 System.Byte[] 类型的本地化资源。
         /// </summary>
-        internal static string HtmlDemoHtml {
+        internal static byte[] HtmlDemoHtml {
             get {
-                return ResourceManager.GetString("HtmlDemoHtml", resourceCulture);
+                object obj = ResourceManager.GetObject("HtmlDemoHtml", resourceCulture);
+                return ((byte[])(obj));
             }
         }
         
         /// <summary>
-        ///   查找类似 function start() {
-        ///    new Vue({
-        ///        el: &apos;#app&apos;,
-        ///        data() {
-        ///            return {
-        ///                info: null,
-        ///                loading: true,
-        ///                errored: false
-        ///            }
-        ///        },
-        ///        filters: {
-        ///            currencydecimal(value) {
-        ///                return value.toFixed(2)
-        ///            }
-        ///        },
-        ///        mounted() {
-        ///            axios
-        ///                .get(&apos;https://api.coindesk.com/v1/bpi/currentprice.json&apos;)
-        ///                .then(response =&gt; {
-        ///                   [字符串的其余部分被截断]&quot;; 的本地化字符串。
+        ///   查找 System.Byte[] 类型的本地化资源。
         /// </summary>
-        internal static string IndexDemoJS {
+        internal static byte[] IndexDemoJS {
             get {
-                return ResourceManager.GetString("IndexDemoJS", resourceCulture);
+                object obj = ResourceManager.GetObject("IndexDemoJS", resourceCulture);
+                return ((byte[])(obj));
             }
         }
     }
