@@ -12,14 +12,14 @@ using System.Security.Cryptography.X509Certificates;
 [DllIN]
 public class NettyDemoCS : INetty
 {
-    private IChannel bootstrapChannel;
+    private IChannel? bootstrapChannel;
     public override void Start(MultithreadEventLoopGroup bossGroup, MultithreadEventLoopGroup workerGroup)
     {
 
     }
     public void Start(MultithreadEventLoopGroup bossGroup, MultithreadEventLoopGroup workerGroup, int port)
     {
-        X509Certificate2 tlsCertificate = null;
+        X509Certificate2? tlsCertificate = null;
         //tlsCertificate = new X509Certificate2(Path.Combine(local, "dotnetty.com.pfx"), "password");
         var bootstrap = new ServerBootstrap();
         bootstrap
@@ -43,7 +43,7 @@ public class NettyDemoCS : INetty
 
     public override void Stop()
     {
-        bootstrapChannel.CloseAsync().Wait();
+        bootstrapChannel?.CloseAsync().Wait();
     }
 }
 
